+++ pkg/common/date/parse.go
@@ -10,6 +10,9 @@ import (
 // ErrDateFormat is returned when there is an error parsing a string date.
 var ErrDateFormat = errors.New("unable to parse date")
 
+// ErrMarshal is returned when there is an error marshaling a time.Time date.
+var ErrMarshal = errors.New("unable to marshal date")
+
 // DefaultFormat is the datetime format used across the provider.
 const DefaultFormat = "2006-01-02T15:04:05Z"
 
@@ -32,4 +35,14 @@ func ParseFormat(format, value string) (time.Time, error) {
 // which is commonly used in the Edge API.
 func FormatRFC3339(t time.Time) string {
 	return t.Format(time.RFC3339)
+}
+
+// ToString returns given date in the string format
+func ToString(value time.Time) (string, error) {
+	bytes, err := value.MarshalText()
+	if err != nil {
+		return "", fmt.Errorf("%w: %s", ErrMarshal, err)
+	}
+
+	return string(bytes), nil
 }