+++ pkg/providers/cloudaccess/cloudaccess_errors.go
@@ -5,6 +5,8 @@ import "errors"
 var (
 	// ErrCloudAccessKey is returned when operation on cloud access key fails
 	ErrCloudAccessKey = errors.New("cloud access key")
+	// ErrCloudAccessKeys is returned when operation on cloud access keys fails
+	ErrCloudAccessKeys = errors.New("cloud access keys")
 	// ErrCloudAccessKeyProperties is returned when operation on cloud access key properties fails
 	ErrCloudAccessKeyProperties = errors.New("cloud access key properties")
 )